requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
google-adk
google-cloud-firestore
google-cloud-storage
google-generativeai
fastapi
uvicorn
streamlit
python-dotenv
pydantic
python-multipart
pillow
python-jose[cryptography]
httpx==3.3.0
passlib[bcrypt]==1.7.4
streamlit==1.28.2
